+++ tests/testthat/test_evaluate.R
@@ -1309,11 +1309,21 @@ test_that("gaussian evaluations are correct in evaluate()",{
                  metrics = "all")
 
   expect_equal(colnames(e1),
-               c("RMSE", "MAE", "r2m", "r2c", "AIC",
-                 "AICc", "BIC", "Predictions",
-                 "Coefficients"))
+               c("RMSE", "MAE", "NRMSE", "RMSEIQR", "RMSESTD", "RAE", "RRSE",
+                 "MAPE", "MSE", "TAE", "TSE", "r2m", "r2c", "AIC", "AICc", "BIC",
+                 "Predictions", "Coefficients"))
+
   expect_equal(e1$RMSE, 16.16881, tolerance = 1e-4)
   expect_equal(e1$MAE, 13.47778, tolerance = 1e-4)
+  expect_equal(e1$NRMSE, 0.2277298, tolerance = 1e-4)
+  expect_equal(e1$RMSEIQR, 0.5774577, tolerance = 1e-4)
+  expect_equal(e1$RMSESTD, 0.8380279, tolerance = 1e-4)
+  expect_equal(e1$RAE, 0.8666762, tolerance = 1e-4)
+  expect_equal(e1$RRSE, 0.8523542, tolerance = 1e-4)
+  expect_equal(e1$MAPE, 0.4836546, tolerance = 1e-4)
+  expect_equal(e1$MSE, 261.4306, tolerance = 1e-4)
+  expect_equal(e1$TAE, 404.3333, tolerance = 1e-4)
+  expect_equal(e1$TSE, 7842.917, tolerance = 1e-4)
   expect_equal(e1$r2m, 0.2665756, tolerance = 1e-4)
   expect_equal(e1$r2c, 0.2665756, tolerance = 1e-4)
   expect_equal(e1$AIC, 258.1214, tolerance = 1e-4)
@@ -1336,8 +1346,8 @@ test_that("gaussian evaluations are correct in evaluate()",{
                  metrics = "all")
 
   expect_equal(colnames(e2),
-               c("RMSE", "MAE", "Predictions"))
-  expect_equal(colnames(e2), c("RMSE","MAE","Predictions"))
+               c("RMSE", "MAE", "NRMSE", "RMSEIQR", "RMSESTD", "RAE", "RRSE",
+                 "MAPE", "MSE", "TAE", "TSE", "Predictions"))
   expect_equal(e2$RMSE, 16.16881, tolerance = 1e-4)
   expect_equal(e2$MAE, 13.47778, tolerance = 1e-4)
   expect_equal(e2$Predictions[[1]][["Target"]], eval_data$score)
@@ -1359,9 +1369,9 @@ test_that("gaussian evaluations are correct in evaluate()",{
                  metrics = "all")
 
   expect_equal(colnames(e3),
-               c("RMSE", "MAE", "r2m", "r2c", "AIC",
-                 "AICc", "BIC", "Predictions",
-                 "Coefficients"))
+               c("RMSE", "MAE", "NRMSE", "RMSEIQR", "RMSESTD", "RAE", "RRSE",
+                 "MAPE", "MSE", "TAE", "TSE", "r2m", "r2c", "AIC", "AICc", "BIC",
+                 "Predictions", "Coefficients"))
 
   eval_data_3 <- dplyr::bind_rows(eval_data %>% dplyr::mutate(fold_ = 1),
                                   eval_data_2 %>% dplyr::mutate(fold_ = 2)) %>%
@@ -1379,9 +1389,9 @@ test_that("gaussian evaluations are correct in evaluate()",{
     dplyr::mutate(fold_ = as.factor(.data$fold_))
 
   expect_equal(colnames(e4),
-               c("fold_","RMSE", "MAE", "r2m", "r2c", "AIC",
-                 "AICc", "BIC", "Predictions",
-                 "Coefficients"))
+               c("fold_", "RMSE", "MAE", "NRMSE", "RMSEIQR", "RMSESTD", "RAE",
+                 "RRSE", "MAPE", "MSE", "TAE", "TSE", "r2m", "r2c", "AIC", "AICc",
+                 "BIC", "Predictions", "Coefficients"))
 
   e1_e3 <- dplyr::bind_rows(e1, e3) %>%
     dplyr::mutate(fold_ = factor(1:2)) %>%
@@ -1528,7 +1538,9 @@ test_that("gaussian evaluations are correct in evaluate()",{
                  metrics = "all",
                  include_predictions = FALSE)
 
-  expect_equal(colnames(e6), c("fold_","RMSE","MAE"))
+  expect_equal(colnames(e6),
+               c("fold_", "RMSE", "MAE", "NRMSE", "RMSEIQR", "RMSESTD",
+                 "RAE", "RRSE", "MAPE", "MSE", "TAE", "TSE"))
 
   # Errors